fix: macOS emit helpers — text_count: *usize → *u32 type mismatch (#189)
mac native build (aarch64-macos):
```
src/renderer/macos.zig:609:65: error: expected type '*usize', found '*u32'
    emitTextInstance(text_buf[0..], &text_count, ...);
note: parameter type declared here
    text_count: *usize,
```

`text_count` 가 mac paint loop 에서 `u32` 로 선언 (line 484: `var text_count: u32 = 0;`).
helper signature 가 `*usize` 였음 → Zig 가 *u32 → *usize coercion 거부.

Linux cross-compile 은 통과했음 — 같은 `usize == u32` (x86_64-linux 64-bit
target 의 usize 는 u64 인데... 왜 통과했지?). 어쨌든 mac native (aarch64-macos
의 usize=u64) 에선 mismatch 검출. \`emitTextInstance\`, \`emitLigatureMatch\`
의 signature 를 \`*u32\` 로 변경해 mac local 변수 type 과 일치.

Linux + Win 의 text_count 도 u32 (mac 과 동일) — signature 변경 영향 없음.

### 회귀 검증

- Linux native + linux target + tests + Win cross-compile 모두 OK.
- mac cross-compile compile 통과 (link 만 SDK 부재로 fail — 사용자 mac
  환경에서 native build 확인 필요).

+++ src/renderer/macos.zig
@@ -1241,7 +1241,7 @@ pub const MetalRenderer = struct {
 /// 의 bearing 으로 glyph 의 cell 안 위치 계산.
 fn emitTextInstance(
     text_buf: []TextInstance,
-    text_count: *usize,
+    text_count: *u32,
     entry: macos_glyph_atlas.AtlasEntry,
     x: usize,
     fy: f32,
@@ -1277,7 +1277,7 @@ fn emitLigatureMatch(
     self: *MetalRenderer,
     encoder: objc.id,
     text_buf: []TextInstance,
-    text_count: *usize,
+    text_count: *u32,
     x: usize,
     count: usize,
     match: mac_font.LigatureMatch,
